Add Network.DeleteIngress() method
This method allows explicit removal of an ingress network along
with its load-balancing endpoint if there are no other endpoints
in the network.  Prior to this, the libnetwork client would have
to manually search for and remove the ingress load balancing
endpoint from an ingress network.  This was, of course, completely
hacky.

Signed-off-by: Chris Telfer <ctelfer@docker.com>